Break stateEditor out into separate test in order to avoid timeouts.
