+++ .travis.yml
@@ -26,6 +26,7 @@ env:
     - RUN_E2E_TESTS_ACCESSIBILITY=true
     - RUN_E2E_TESTS_COLLECTIONS=true
     - RUN_E2E_TESTS_EDITOR_AND_PLAYER=true
+    - RUN_E2E_TESTS_STATE_EDITOR=true
     - RUN_E2E_TESTS_EDITOR_FEATURES=true
     - RUN_E2E_TESTS_EMBEDDING=true
     - RUN_E2E_TESTS_EXTENSIONS=true
@@ -88,6 +89,7 @@ script:
 - if [ $RUN_E2E_TESTS_ACCESSIBILITY == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="accessibility" --prod_env; fi
 - if [ $RUN_E2E_TESTS_COLLECTIONS == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="collections" --prod_env; fi
 - if [ $RUN_E2E_TESTS_EDITOR_AND_PLAYER == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="editorAndPlayer" --prod_env; fi
+- if [ $RUN_E2E_TESTS_STATE_EDITOR == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="stateEditor" --prod_env; fi
 - if [ $RUN_E2E_TESTS_EDITOR_FEATURES == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="editorFeatures" --prod_env; fi
 - if [ $RUN_E2E_TESTS_EMBEDDING == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="embedding" --prod_env; fi
 - if [ $RUN_E2E_TESTS_EXTENSIONS == 'true' ]; then travis_retry bash scripts/run_e2e_tests.sh --suite="extensions" --prod_env; fi
